+++ src/game/g_items.c
@@ -565,7 +565,7 @@ int Pickup_Weapon( gentity_t *ent, gentity_t *other ) {
 				if ( !( ent->parent->client->PCSpecialPickedUpCount % LT_SPECIAL_PICKUP_MOD ) ) {
 					AddScore( ent->parent, WOLF_AMMO_UP );
 					if ( ent->parent && ent->parent->client ) {
-						G_LogPrintf( "Ammo_Pack: %d %d\n", ent->parent - g_entities, other - g_entities );  // OSP
+						G_LogPrintf( "Ammo_Pack: %d %d\n", (int)(ent->parent - g_entities), (int)(other - g_entities) );  // OSP
 					}
 				}
 				ent->parent->client->PCSpecialPickedUpCount++;
@@ -676,7 +676,7 @@ int Pickup_Health( gentity_t *ent, gentity_t *other ) {
 		if ( ent->parent && ent->parent->client && other->client->sess.sessionTeam == ent->parent->client->sess.sessionTeam ) {
 			if ( !( ent->parent->client->PCSpecialPickedUpCount % MEDIC_SPECIAL_PICKUP_MOD ) ) {
 				AddScore( ent->parent, WOLF_HEALTH_UP );
-				G_LogPrintf( "Health_Pack: %d %d\n", ent->parent - g_entities, other - g_entities );    // OSP
+				G_LogPrintf( "Health_Pack: %d %d\n", (int)(ent->parent - g_entities), (int)(other - g_entities) );    // OSP
 			}
 			G_AddSkillPoints( ent->parent, SK_FIRST_AID, 1.f );
 			G_DebugAddSkillPoints( ent->parent, SK_FIRST_AID, 1.f, "health pack picked up" );
