+++ src/game/g_match.c
@@ -437,7 +437,7 @@ const char *G_createStats( gentity_t *refEnt ) {
 		}
 	}
 
-	return( va( "%d %d %d%s %d%s", refEnt - g_entities,
+	return( va( "%d %d %d%s %d%s", (int)(refEnt - g_entities),
 				refEnt->client->sess.rounds,
 				dwWeaponMask,
 				strWeapInfo,
